+++ src/api/resources/sitemap/index.js
@@ -4,12 +4,19 @@ import type { Server } from 'restify';
 
 import _ from 'lodash';
 
+const sendXml = (res, data) => {
+  res.writeHead(200, {
+    'Content-Type': 'application/xml',
+  });
+  res.write(data);
+  res.end();
+};
+
 export default function SitemapResource (server: Server, controller: any) {
   server.get('/sitemap.xml', async (req, res, next) => {
     try {
-      const sitemapIndex = await controller.index();
-      res.setHeader('Content-Type', 'application/xml');
-      res.send(200, sitemapIndex);
+      const index = await controller.index();
+      sendXml(res, index, next);
     } catch (error) {
       console.error(error);
       res.send(500);
@@ -22,9 +29,8 @@ export default function SitemapResource (server: Server, controller: any) {
     try {
       // eslint-disable-next-line quote-props
       const { '0': resource, '1': page } = req.params;
-      const sitemapIndex = await controller.generate(resource, _.toInteger(page));
-      res.setHeader('Content-Type', 'application/xml');
-      res.send(200, sitemapIndex);
+      const sitemap = await controller.generate(resource, _.toInteger(page));
+      sendXml(res, sitemap, next);
     } catch (error) {
       console.error(error);
       res.send(500);